# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contoh API documentation",
        "description": "Contoh API doc menggunakan OpenAPI/Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "ibanezcentivoliaahista@mail.ugm.ac.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/api/gallery": {
            "post": {
                "tags": [
                    "gallery"
                ],
                "summary": "Upload gallery image",
                "description": "Upload an image along with title and description",
                "operationId": "uploadGallery",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "file"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "judul",
                                        "type": "string",
                                        "example": "Bulan dan Bintang"
                                    },
                                    "description": {
                                        "description": "deskripsi",
                                        "type": "string",
                                        "example": "Pertemuan Pertama Bulan dan Bintang"
                                    },
                                    "file": {
                                        "description": "gambar",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil menambahkan gambar",
                                    "data": {
                                        "output": "halo nama file deskripsi gambar",
                                        "title": "judul",
                                        "description": "alamak",
                                        "file": "gambar"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/greet": {
            "get": {
                "tags": [
                    "greeting"
                ],
                "summary": "Returns a Sample API response",
                "description": "A sample greeting to test out the API",
                "operationId": "greet",
                "parameters": [
                    {
                        "name": "firstname",
                        "in": "query",
                        "description": "nama depan",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "description": "nama belakang",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil mengambil Kategori Berita",
                                    "data": {
                                        "output": "Hallo John Doe",
                                        "firstname": "John",
                                        "lastname": "Doe"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "gallery",
            "description": "gallery"
        },
        {
            "name": "greeting",
            "description": "greeting"
        }
    ]
}